pyproject.toml
[project]
name = "fcan"
version = "0.1.0"
description = "Function Calling A2A Network: a framework that empowers agents in A2A networks with function calling"
readme = "readme.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "ollama>=0.4.8",
    "python-ulid>=3.0.0",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["source/fcan"]
